cryptor: An R wrapper for the CryptoCompare API.

This package provides access to the crypto currency API documented at https://www.cryptocompare.com/api. The API is currently public and can be accessed without authentication or credentials. The API authors have enforced hourly and second call limits.

Functions

get_api_limit - Get API calls made and remaining against current limits

get_coins - List all available coins

get_price - Get price for a currency or group of currencies

get_price_details - Get price with additional details for a currency or group of currencies

get_historical_price - Get price for a currency from a specifc date / timestamp

get_social - Get comprehensive social data for a given currency